make derived leaf types generic over Field, Bool
